+++ pkg/workflow/controllers/chaos_node_test.go
@@ -59,7 +59,7 @@ var _ = Describe("Workflow", func() {
 		It("could spawn one chaos", func() {
 			ctx := context.TODO()
 			now := time.Now()
-			duration := 5 * time.Second
+			duration := 30 * time.Second
 
 			By("create simple chaos node with pod chaos")
 			startTime := metav1.NewTime(now)
@@ -106,7 +106,7 @@ var _ = Describe("Workflow", func() {
 		It("could spawn one schedule", func() {
 			ctx := context.TODO()
 			now := time.Now()
-			duration := 5 * time.Second
+			duration := 30 * time.Second
 
 			By("create simple chaos node with schedule")
 			startTime := metav1.NewTime(now)